Refactor 'utils' into 'helpers' folder
With this refactorization things that changed includes:
- Require lines updated to new directory
- Creation of new 'helpers' folder
- Deletion of existing 'utils' folder
