+++ helpers/showroomUtil.js
@@ -1,6 +1,6 @@
 const Discord = require('discord.js');
 const fetch = require('node-fetch');
-const cl = require('../utils/crystaliaLibrary.js')
+const cl = require('../helpers/crystaliaLibrary.js')
 const { JSDOM } = require('jsdom');
 
 const prefix = process.env.prefix;
